fix: rename package from util to utilities and update dependencies
