+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
   - repo: https://github.com/gitleaks/gitleaks
-    rev: v8.25.1
+    rev: v8.27.2
     hooks:
       - id: gitleaks
   - repo: https://github.com/Yelp/detect-secrets
