+++ github/repo.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/google/go-github/v72/github"
+	"github.com/google/go-github/v73/github"
 )
 
 // ListRepositories retrieves repositories for the given target (user or organization).
