+++ github/types.go
@@ -1,7 +1,7 @@
 package github
 
 import (
-	"github.com/google/go-github/v72/github"
+	"github.com/google/go-github/v73/github"
 )
 
 type Service struct {
