+++ renovate.json
@@ -3,30 +3,30 @@
   "baseBranches": [
     "main"
   ],
+  "commitBody": "Generated by renovateBot",
   "extends": [
     "config:recommended"
   ],
-  "postUpdateOptions": [
-    "gomodTidy"
-  ],
-  "commitBody": "Generated by renovateBot",
+  "ignorePaths": [],
   "packageRules": [
     {
-      "schedule": "before 6am",
-      "timezone": "Europe/Berlin",
+      "groupName": "github workflows",
       "matchManagers": [
         "github-actions"
       ],
-      "groupName": "github workflows"
+      "schedule": "before 6am",
+      "timezone": "Europe/Berlin"
     },
     {
-      "schedule": "before 6am",
-      "timezone": "Europe/Berlin",
+      "groupName": "go deps",
       "matchManagers": [
         "gomod"
       ],
-      "groupName": "go deps"
+      "schedule": "before 6am",
+      "timezone": "Europe/Berlin"
     }
   ],
-  "ignorePaths": []
+  "postUpdateOptions": [
+    "gomodTidy"
+  ]
 }
