--- util/docs.go
@@ -1,2 +1,0 @@
-// Package util provides utility functions.
-package util
