+++ utilities/dir.go
@@ -1,4 +1,4 @@
-package util
+package utilities
 
 import (
 	"os"
