+++ utilities/docs.go
@@ -1,0 +1,2 @@
+// Package utilities provides utility functions.
+package utilities
